section 13 14

+++ TodoApp/main.py
@@ -1,4 +1,5 @@
 
+from pyexpat import model
 from typing import Optional
 from xmlrpc.client import boolean
 from fastapi import FastAPI, Depends, HTTPException
@@ -8,6 +9,8 @@ from database import engine, SessionLocal
 from sqlalchemy.orm import Session
 
 from pydantic import BaseModel, Field
+
+from auth import get_user_exceptions, get_current_user
 
 app = FastAPI()
 
@@ -45,9 +48,11 @@ async def read_all(db: Session = Depends(get_db)):
 
 
 @app.get("/todo/{todo_id}")
-async def read_todo(todo_id: int, db: Session = Depends(get_db)):
+async def read_todo(todo_id: int, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exceptions()
     todo_record = db.query(models.Todos).filter(
-        models.Todos.id == todo_id).first()
+        models.Todos.id == todo_id).filter(models.Todos.owner_id == user.get("id")).first()
     if todo_record is not None:
         return todo_record
     else:
@@ -56,23 +61,28 @@ async def read_todo(todo_id: int, db: Session = Depends(get_db)):
 
 # add a todo model record in db
 @app.post("/")
-async def create_todo(todo: Todo, db: Session = Depends(get_db)):
+async def create_todo(todo: Todo, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exceptions()
     todo_model = models.Todos()
     todo_model.title = todo.title
     todo_model.description = todo.description
     todo_model.priority = todo.priority
     todo_model.complete = todo.complete
+    todo_model.owner_id = user.get("id")
     db.add(todo_model)  # add a todo_model object to session
     db.commit()  # flush pending changes in the session to database and commit the transaction
     return successful_response(201)
 
 
 @app.put("/todo/{todo_id}")
-async def update_todo(todo_id: int, todo: Todo, db: Session = Depends(get_db)):
+async def update_todo(todo_id: int, todo: Todo, user: dict = Depends(get_current_user),  db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exceptions()
+
     todo_model = db.query(models.Todos)\
         .filter(
-        models.Todos.id == todo_id)\
-        .first()
+        models.Todos.id == todo_id).filter(models.Todos.owner_id == user.get("id")).first()
     if todo_model is None:
         raise http_exception()
 
@@ -86,14 +96,26 @@ async def update_todo(todo_id: int, todo: Todo, db: Session = Depends(get_db)):
 
 
 @app.delete("/todo/{todo_id}")
-async def delete_todo(todo_id: int, db: Session = Depends(get_db)):
+async def delete_todo(todo_id: int, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+
+    if user is None:
+        raise get_user_exceptions()
+
     todo_model = db.query(models.Todos).filter(
-        models.Todos.id == todo_id).first()
+        models.Todos.id == todo_id).filter(models.Todos.owner_id == user.get("id")).first()
     if todo_model is None:
         raise http_exception()
     db.query(models.Todos).filter(models.Todos.id == todo_id).delete()
     db.commit()
     return successful_response(200)
+
+
+@app.get("/todos/user")
+async def read_all_by_user(user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exceptions()
+    print(user)
+    return db.query(models.Todos).filter(models.Todos.owner_id == user.get("id")).all()
 
 
 def successful_response(status_code: int):
